% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_breaks_to_labels.R
\name{age_breaks_to_labels}
\alias{age_breaks_to_labels}
\title{Convert Age Break String to Age Group Labels}
\usage{
age_breaks_to_labels(age_string)
}
\arguments{
\item{age_string}{A single character string of age breakpoints, separated by semicolons.
Must include at least two breakpoints. The last breakpoint can be `Inf` to indicate
an open-ended upper age group.}
}
\value{
A character vector of formatted age group labels.
}
\description{
Converts a semicolon-separated string of numeric age breakpoints into a character
vector of age group labels. For example, the input `"0;5;10;...;80;Inf"` will produce
the labels `"0–4"`, `"5–9"`, ..., `"80+"`.
}
\examples{
age_breaks_to_labels("0;5;10;15;20;25;30;35;40;45;50;55;60;65;70;75;80;Inf")
# Returns:
# [1] "0–4"   "5–9"   "10–14" "15–19" "20–24" ... "75–79" "80+"

}
