% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_inputs.R
\name{aggregate_inputs}
\alias{aggregate_inputs}
\title{Aggregate or Default Model Inputs by Age Structure}
\usage{
aggregate_inputs(
  preprocessed,
  cv_params,
  seed_data,
  new_age_breaks,
  aggregate_age,
  age_vaccination_beta_modifier
)
}
\arguments{
\item{preprocessed}{A list from prepare_model_inputs() including demographic data.}

\item{cv_params}{Output from case_vaccine_to_param().}

\item{seed_data}{Data frame of seeding values (e.g. for WHO).}

\item{new_age_breaks}{Numeric vector of age breakpoints.}

\item{aggregate_age}{Logical. If TRUE, aggregates into new_age_breaks.}

\item{age_vaccination_beta_modifier}{A data frame of protection modifiers by dose and age.}
}
\value{
A list with two elements:
  \itemize{
    \item default_inputs: Model inputs without age aggregation.
    \item inputs: Aggregated model inputs (if aggregate_age is TRUE), else same as default.
  }
}
\description{
This helper function performs optional aggregation of model inputs into a new age structure.
It returns both the default unaggregated inputs and the aggregated inputs (if requested).
}
\keyword{internal}
