% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{contact_matricies}
\alias{contact_matricies}
\title{Social Contact Matrices from Prem et al. (2017)}
\format{
A list of 152 numeric 16x16 matrices.
}
\source{
Prem K, Cook AR, Jit M (2017). PLoS Comput Biol 13(9): e1005697. https://doi.org/10.1371/journal.pcbi.1005697
}
\usage{
contact_matricies
}
\description{
These are the contact matrices generated from the paper "Projecting social contact matrices in 152 countries using contact surveys and demographic data" by Prem et al., (2017). Each matrix estimates contact rates across 16 age groups (0–4, 5–9, ..., 75–79) for 152 countries. Data is found in a list where the name of the object in the list is the country/territories ISO3 code.
}
\keyword{datasets}
