% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_vaccination_data.R
\name{expand_pre1980_vaccination}
\alias{expand_pre1980_vaccination}
\title{Expand Pre-1980 Routine Vaccination Coverage}
\usage{
expand_pre1980_vaccination(
  processed_vaccination,
  vaccination_pre1980,
  disease,
  iso
)
}
\arguments{
\item{processed_vaccination}{Routine coverage data (to determine first year and vaccine).}

\item{vaccination_pre1980}{Data frame of historical vaccine assumptions (intro year, starting coverage).}

\item{disease}{Disease name (lowercase) to match.}

\item{iso}{3-letter ISO country code.}
}
\value{
A data frame of additional rows to prepend to coverage time series.
}
\description{
Uses historical assumptions to estimate vaccination coverage prior to the first observed year.
Applies linear interpolation from an assumed introduction year up to the first observed coverage year.
}
\keyword{internal}
