% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helpers.R
\name{get_years}
\alias{get_years}
\title{Generate a Sequence of Years from a Vector and Optional Start/End Bounds}
\usage{
get_years(yrs, start, end)
}
\arguments{
\item{yrs}{A numeric vector of years from which to derive default start or end values.}

\item{start}{A numeric or character value specifying the start year. If `""`, the minimum year in `yrs` is used.}

\item{end}{A numeric or character value specifying the end year. If `""`, the maximum year in `yrs` is used.}
}
\value{
A numeric vector of consecutive years from `start` to `end` (inclusive).
}
\description{
This helper function returns a sequence of years based on a full vector of years (`yrs`)
and optional start and end bounds. If either `start` or `end` is an empty string,
the minimum or maximum value in `yrs` is used, respectively.
}
\examples{
get_years(2000:2020, start = 2005, end = 2010)
get_years(2000:2020, start = "", end = 2010)
get_years(2000:2020, start = 2005, end = "")
get_years(2000:2020, start = "", end = "")

}
\keyword{internal}
