% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helpers.R
\name{reformat_contact_matrix}
\alias{reformat_contact_matrix}
\title{Reformat Contact Matrix to Match Age Vector}
\usage{
reformat_contact_matrix(contact_matrix_raw, age_vector)
}
\arguments{
\item{contact_matrix_raw}{A square matrix of contact rates, with rows and columns corresponding to age groups in 5-year intervals.}

\item{age_vector}{A vector of age categories that the contact matrix will be reformatted to match.}
}
\value{
A matrix of contact rates with dimensions matching the provided age vector.
}
\description{
This function reformats a raw contact matrix (which uses age groups in 5-year intervals)
to match a given age vector, mapping each age group in the contact matrix to the nearest
value in the age vector.
}
\keyword{internal}
