% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{UN_WPP_fertility}
\alias{UN_WPP_fertility}
\title{UN World Population Prospects 2024: Fertility Rates}
\format{
A data frame with columns:
- area: Country or territory
- iso3: 3-digit ISO code
- year: Year
- x15 to x49: Live births per 1,000 women, by age
}
\source{
UN DESA Population Division, World Population Prospects 2024
}
\usage{
UN_WPP_fertility
}
\description{
Estimated Total Fertility Rate (live births per 1,000 women) by age (15–49) and country, from the 2024 UN WPP.
}
\keyword{datasets}
