% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{UN_WPP_migration}
\alias{UN_WPP_migration}
\title{UN World Population Prospects 2024: Net Migration Rate}
\format{
A data frame with columns:
- area: Country or territory
- iso3: 3-digit ISO code
- year: Year
- migration_rate_1000: Net migration rate
}
\source{
UN DESA Population Division, World Population Prospects 2024
}
\usage{
UN_WPP_migration
}
\description{
Net migration rate per 1,000 population, with no age disaggregation, from the 2024 UN WPP.
}
\keyword{datasets}
