% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{UN_WPP_mortality}
\alias{UN_WPP_mortality}
\title{UN World Population Prospects 2024: Crude Death Rates}
\format{
A data frame with columns:
- area: Country or territory
- iso3: 3-digit ISO code
- year: Year
- x0 to x100: Crude death rates
}
\source{
UN DESA Population Division, World Population Prospects 2024
}
\usage{
UN_WPP_mortality
}
\description{
Crude Death Rates (deaths per 1,000 population) by single year age bands 0–100, from the 2024 UN WPP.
}
\keyword{datasets}
