% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{UN_WPP_population_total}
\alias{UN_WPP_population_total}
\title{UN World Population Prospects 2024: Total Population}
\format{
A data frame with columns:
- area: Country or territory
- iso3: 3-digit ISO code
- year: Year
- x0 to x100: Total population counts
}
\source{
UN DESA Population Division, World Population Prospects 2024
}
\usage{
UN_WPP_population_total
}
\description{
Total population (in thousands) by single year of age (0–100), as of 1 July, from the 2024 UN WPP.
}
\keyword{datasets}
