% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_dust2.R
\name{unpack_dust2}
\alias{unpack_dust2}
\title{Unpack Dust Model State Data}
\usage{
unpack_dust2(
  model_system,
  model_object,
  dimension_names,
  which_state_dimensions
)
}
\arguments{
\item{model_system}{The model system object containing packing information and particle details.}

\item{model_object}{The dust model object containing state data.}

\item{dimension_names}{List of dimension names for states, time, and other variables.}

\item{which_state_dimensions}{List mapping state names to their associated dimensions.}
}
\value{
A data frame containing the unpacked dust model state data.
}
\description{
This function extracts and unpacks data from a dust model state object, separating compartments and non-compartments.
}
