% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{VIMC_vaccination_sia}
\alias{VIMC_vaccination_sia}
\title{Supplementary Immunization Activities (SIA) Coverage Estimates}
\format{
A data frame with columns:
- area: Country
- year: Year
- age: Age
- coverage: Coverage
- disease: Disease
- vaccine: Vaccine
}
\source{
Shattock et al., 2024, The Lancet. https://doi.org/10.1016/S0140-6736(24)00850-X
}
\usage{
VIMC_vaccination_sia
}
\description{
Coverage estimates for SIAs from WHO and supplementary sources, as used in Shattock et al. (2024).
}
\keyword{datasets}
