% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{WHO_disease_reports}
\alias{WHO_disease_reports}
\title{WHO Global Reported Cases of Vaccine-Preventable Diseases (1980–2023)}
\format{
A data frame with columns:
- iso3: 3-digit ISO country code
- name: Country or territory name
- year: Year of report
- disease_short: Short disease name
- disease_description: Full description of disease
- cases: Number of reported cases
}
\source{
WHO Immunization Dashboard: https://immunizationdata.who.int/
}
\usage{
WHO_disease_reports
}
\description{
Annual country-level case reports for vaccine-preventable diseases collected by WHO via the Joint Reporting Form.
}
\keyword{datasets}
