% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{WHO_vaccination_routine}
\alias{WHO_vaccination_routine}
\title{WHO Routine Infant Immunization Coverage (WUENIC)}
\format{
A data frame with columns:
- iso3: 3-digit ISO code
- area: Country or territory
- vaccine: Vaccine name
- vaccine_description: Description of vaccine
- coverage_category: Type of coverage
- coverage_description: Coverage description
- coverage: Coverage percent
- dose_order: Order of dose
- disease: Disease
}
\source{
WHO Immunization Dashboard: https://immunizationdata.who.int/global
}
\usage{
WHO_vaccination_routine
}
\description{
Routine immunization coverage estimates (WUENIC) by country, vaccine, dose, and year.
}
\keyword{datasets}
