% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{WHO_vaccination_schedule}
\alias{WHO_vaccination_schedule}
\title{WHO Reported Vaccination Schedules}
\format{
A data frame with columns:
- iso3: 3-digit ISO code
- area: Country or territory
- WHO_region: WHO region
- year: Year
- vaccine_code: Vaccine identifier code
- vaccine_description: Vaccine description
- schedulerounds: Schedule rounds
- target_pop: Target population
- target_pop_description: Description of target group
- geoarea: Geographic area
- age_administered: Age at administration
- comment: Comments
}
\source{
WHO Immunization Dashboard: https://immunizationdata.who.int/global
}
\usage{
WHO_vaccination_schedule
}
\description{
Current vaccination schedules for children, adolescents, and adults as reported to the WHO.
}
\keyword{datasets}
