% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_unpack_results.R
\name{run_model_unpack_results}
\alias{run_model_unpack_results}
\title{Run and Unpack Dust Model Simulation Results}
\usage{
run_model_unpack_results(odin_model, params, time = 1000, no_runs = 10)
}
\arguments{
\item{odin_model}{An unevaluated odin model function used to construct the dust system.}

\item{params}{A named list of model parameters including `n_age`, `n_vacc`, `n_risk`, and `tt_migration`.}

\item{time}{Integer. Total number of time steps to simulate (default is 1000). Time starts at 0.}

\item{no_runs}{Integer. Number of stochastic particles to simulate (default is 10).}
}
\value{
A `data.table` in long format containing simulation results by state, time, and strata.
}
\description{
Initializes and runs a Dust model using a provided odin model and parameter set, then unpacks the simulation
output into a long-format `data.table` using `unpack_dust2()`. Suitable for analysis and plotting.
}
\details{
The function uses `dust2::dust_system_create()` to instantiate and run the simulation, and `unpack_dust2()`
to reshape the result. The output includes states such as `S`, `E`, `I`, `R`, `Is`, `Rc`, `new_case`, `Reff_age`, etc.
}
